Convert `brew missing` test to spec.
